iProject2
Created a memory display for index.html and a few more global variables.
Utils add a function to help display the memory to the user. The
following classes were created: mainMemory, memoryManagement, and pcb.
This update refers to the load command and allowed a user to enter a
error free program and have it load into memory and is shown in the
memory display.