major updates to CPU and program can run
updates in cpu fix each opcode and what it does. Should be mostly
correct now. programs will sync with clock pulse and  added system call
opcode to help print out the output of the programs and also a EA opcode
that just increases the PC. Added a system break op code 00 which allows
the program to stop and update. Added EE which will increment a byte in
from memory. Added real time updates as well as program executes.