Clean up
Cleaned up some code to make project 3 a bit easier.  added a couple
globals that will reference memory partitions,  and RR quantum, as well
as another array to keep track of active programs, maybe not necessary.
Added a wider view for memory to see all partitions in CoreMemory.js
Fixed a lot of confusing code in CPU to make it known which pcb is
running and when, changed shell to reflect the cpu and pcb changes