More clean up and setup
edited the memory partition locations in globals.js to reflect their
actual values and also created a quantum var and also a var to dictate
which schedule to run.  Temporarily turned off glados in index.js Don't
worry I will hopefully remember to turn it back on by Thursday. Fixed  a
bug in cpu.js that would print the correct letters (it was a parsing bug
with the memory address FYI) and added a kill feature that needs more
testing.  Added a func to find the start of a program that the PCBs use.
hopefully will be useful with the kill function and runall. Started the
cpuSchedule.js class which obviously need some more work.