Runall function added and some memory checking
Added the function to Runall programs in the ready queue and it is
running smooooooth. Also added a memory bounds check to make sure that
programs are behaving (oh behave!) Also made sure to remove the programs
from the readyqueue after they have been run. next move would be to let
the memory know that slot can be reopen for business and other things.