More Functionality
Added the ability to view the currently running process PCB in a table
below the canvas. the CPU is updated to help control that table of
insertions of current processes and deletions of finished processes. So
far only one will be shown at a time until i get RR up and running. The
shell has an added kill command to kill a currently running process.