Big Update
RR is functional updated a lot in the cpu to make that happen. Added a
context switch function to cpuScheduler and did a small edit to the
console to allow Glados (which is now uncommented, told you I would
remember) to be more accepting with scrolling. Took out the middleman in
cpu.js just replaced all this.pcb with _currentPCB and it made a big
difference in readability.