updates with swapping
fixed some bugs with the program display below the canvas. added a
location field to the table so we can track where the program is. FCFS
is added and works, will work on priority later(you could say it's not a
priority right now :)  ) . Will start to work on grabbing the 4th
program from the disk and implementing that.