finals
fixed a bug in the filesystem that wasnt reading the file as a string
when doing a replace to get rid of the '-' filller. Updated memory to
handle the disk swapping better when the 4th program is loaded. added a
var to pcb that allowed me to keep track of whether it is on disk or
not.